[Management] Modifications of br-ex script (NOT TESTED)
The script which aims at creating and configuring br-ex has been
modified.
In a few words, since we are tuning network configuration over SSH,
all the commands to execute must be sent in one go (the resulting code
is pretty ugly).

It might work, but I have not tested it yet since the network
configuration of c44 is broken due to a previous experiment
(my bad x_X).